Moved loernwerk error to own file to avoid importing typeorm in frontend
