+++ backend/controller/AccountController.ts
@@ -2,7 +2,7 @@ import { DBUser } from '../../model/user/DBUser';
 import { IUser, UserClass } from '../../model/user/IUser';
 import bcrypt from 'bcrypt';
 import crypto from 'crypto';
-import { LoernwerkError, LoernwerkErrorCodes } from '../loernwerkUtilities';
+import { LoernwerkError, LoernwerkErrorCodes } from '../loernwerkError';
 import { DBSequence } from '../../model/sequence/DBSequence';
 import { SequenceController } from '../controller/SequenceController';
 /**
